+++ tests/testthat/test-interp2.R
@@ -19,6 +19,9 @@ test_that("interp2: output accuracy", {
   random_vals <- sample(2:(2^10), 10)
   for (i in random_vals) {
     curr_mult <- runif(n = 1, min = 1, max = i)
-    expect_equal((interp2(c(i, 2 * i), c(0, i) * curr_mult, i + 1)), (0:i) * curr_mult)
+    expect_equal(
+      (interp2(c(i, 2 * i), c(0, i) * curr_mult, i + 1)),
+      (0:i) * curr_mult
+      )
   }
 })