+++ NEWS.md
@@ -1,3 +1,16 @@
-# mousetRajectory (development version)
+# mousetRajectory 0.2.1
 
 * Initial CRAN submission.
+
+---
+
+# mousetRajectory 0.2.0
+
+* 'time_circle_entered' and 'time_circle_left' arguments have been converted to snake case ('xMid' -> 'x_mid' and 'yMid' -> 'y_mid')
+
+---
+
+# mousetRajectory 0.1.0 
+
+* wuelib has been re-named into mousetRajectory! Yeah!
+